requirements.txt
Django~=4.2.6
django-filter~=23.3
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
djoser~=2.2.0
gunicorn~=21.2.0
python-dotenv~=1.0.0
redis~=5.0.1
# "python-socketio[client]"~=5.8.0
psycopg2-binary>=2.9
